requirements.txt
Flask_MySQLdb==0.2.0
Flask==1.1.2
passlib==1.7.2
mysqlclient==1.4.6
